% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_bernoulli.R
\name{conditional_bernoulli}
\alias{conditional_bernoulli}
\title{conditional bernoulli distribution}
\usage{
conditional_bernoulli(p, psi, dim = 1)
}
\arguments{
\item{p}{matrix (of dimension \code{dim} x K) of (conditional) probabilities
of success}

\item{psi}{scalar or column vector (of length \code{dim}) of probabilities
for the latent bernoulli variable}

\item{dim}{a scalar giving the number of rows in the resulting greta array}
}
\description{
greta probability distribution over a K-dimensional vector of
  binary variables, arising from independent Bernoulli draws each conditioned
  on a single draw from another Bernoulli draw.
}
\details{
A compound distribution, where elements of the bernoulli vector
  variable \emph{y} can only be 1 if a scalar latent bernoulli variable
  \emph{z} takes value 1, i.e.:

  \deqn{ y_i ~ bernoulli(z * p_i)}
  \deqn{z ~ bernoulli(psi)}
  where
  \deqn{p_i = p(y_i = 1 | z = 1)}
  \deqn{psi = p(z = 1)}

  \emph{p} and \emph{psi} are distinguishable provided there are multiple
  trials in each observation of \emph{y}. The density of this compound
  distribution can be calculated directly, explicitly integrating over the
  latent variable \emph{z}, as:

  \deqn{psi * prod((p ^ y) * (1 - p) ^ (1 - y)) + max(y) * (1 - psi)}

  This formulation underpins the ecological imperfect-detection model of
  MacKenzie et al. where \emph{y} and \emph{p} are vectors indicating whether
  a species was detected at each visit, and the probability of detection
  (which may vary between visits), and \emph{z} and \emph{psi} are scalars
  indicating whether the species was present (assumed to be the same at all
  visits) and the probability of being present.
}
\references{
MacKenzie, D. I., Nichols, J. D., Lachman, G. B., Droege, S.,
  Andrew Royle, J., & Langtimm, C. A. (2002). Estimating site occupancy rates
  when detection probabilities are less than one. \emph{Ecology}, 83(8),
  2248-2255.
}
