% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{gretaMultivariate}
\alias{gretaMultivariate}
\alias{gretaMultivariate-package}
\title{Multivariate Modelling with greta}
\description{
A greta extension providing distributions and functions for
multivariate modelling (modelling multivariate response variables).
}
