% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_probit.R
\name{multivariate_probit}
\alias{multivariate_probit}
\title{multivariate probit distribution}
\usage{
multivariate_probit(mean, C, dim = 1)
}
\arguments{
\item{mean}{matrix or column vector (of length K) of unconstrained parameters
giving the mean of the latent normal parameters}

\item{C}{K-dimensional correlation matrix}

\item{dim}{a scalar giving the number of rows in the resulting greta array}
}
\description{
greta probability distribution over a K-dimensional vector of
  binary variables.
}
