% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_inflated_poisson.R
\name{zero_inflated_poisson}
\alias{zero_inflated_poisson}
\title{Zero inflated (poisson) distribution}
\usage{
zero_inflated_poisson(theta, lambda, dim = 1)
}
\arguments{
\item{theta, }{scalar for the zero-inflation parameter}

\item{lambda, }{scalar the expected value of the poisson distribution}

\item{dim}{a scalar giving the number of rows in the resulting greta array}
}
\description{
greta probability distribution for a zero-inflated poisson distribution
}
\details{
Zero inflated distribution with:
     if(y = 0)
     \deqn{theta + (1 - theta) * Poisson(lambda)}
     else
     \deqn{(1 - theta) * Poisson(lambda)}
}
